Add input sanitization
