+++ packages/gafl-webapp-service/src/handlers/attribution-handler.js
@@ -1,7 +1,5 @@
 import { UTM, ATTRIBUTION_REDIRECT_DEFAULT } from '../constants.js'
 
-const sanitiseInput = value => (value ? encodeURIComponent(value) : value)
-
 /**
  * Attribution route handler
  * @param request
@@ -13,11 +11,11 @@ export default async (request, h) => {
   const cache = request.cache()
   await cache.helpers.status.set({
     attribution: {
-      [UTM.CAMPAIGN]: sanitiseInput(request.query[UTM.CAMPAIGN]),
-      [UTM.MEDIUM]: sanitiseInput(request.query[UTM.MEDIUM]),
-      [UTM.CONTENT]: sanitiseInput(request.query[UTM.CONTENT]),
-      [UTM.SOURCE]: sanitiseInput(request.query[UTM.SOURCE]),
-      [UTM.TERM]: sanitiseInput(request.query[UTM.TERM])
+      [UTM.CAMPAIGN]: request.query[UTM.CAMPAIGN],
+      [UTM.MEDIUM]: request.query[UTM.MEDIUM],
+      [UTM.CONTENT]: request.query[UTM.CONTENT],
+      [UTM.SOURCE]: request.query[UTM.SOURCE],
+      [UTM.TERM]: request.query[UTM.TERM]
     }
   })
   return h.redirect(redirectTarget)
