+++ packages/gafl-webapp-service/src/server.js
@@ -4,31 +4,25 @@
 
 import Hapi from '@hapi/hapi'
 import CatboxRedis from '@hapi/catbox-redis'
-import Vision from '@hapi/vision'
-import Inert from '@hapi/inert'
-import Scooter from '@hapi/scooter'
-import HapiGapi from '@defra/hapi-gapi'
-import Crumb from '@hapi/crumb'
-import Blankie from 'blankie'
 import Nunjucks from 'nunjucks'
 import find from 'find'
 import path from 'path'
 import Dirname from '../dirname.cjs'
 import routes from './routes/routes.js'
 import {
+  CHANNEL_DEFAULT,
   CSRF_TOKEN_COOKIE_NAME_DEFAULT,
+  FEEDBACK_URI_DEFAULT,
   REDIS_PORT_DEFAULT,
   SESSION_COOKIE_NAME_DEFAULT,
-  SESSION_TTL_MS_DEFAULT,
-  FEEDBACK_URI_DEFAULT,
-  UTM,
-  CHANNEL_DEFAULT
+  SESSION_TTL_MS_DEFAULT
 } from './constants.js'
-import { COOKIES, REFUND_POLICY, ACCESSIBILITY_STATEMENT, PRIVACY_POLICY } from './uri.js'
+import { ACCESSIBILITY_STATEMENT, COOKIES, PRIVACY_POLICY, REFUND_POLICY } from './uri.js'
 
-import sessionManager, { useSessionCookie } from './session-cache/session-manager.js'
+import sessionManager from './session-cache/session-manager.js'
 import { cacheDecorator } from './session-cache/cache-decorator.js'
 import { errorHandler } from './handlers/error-handler.js'
+import { getPlugIns } from './plugins.js'
 
 let server
 
@@ -63,83 +57,8 @@ const createServer = options => {
 /*
  * The hapi plugins and their options which will be registered on initialization
  */
-
-// This is a hash of the inline script at line 31 of the GDS template. It is added to the CSP to except the in-line
-// script. It needs the quotes.
-const scriptHash = "'sha256-+6WnXIl4mbFTCARd8N3COQmT3bJJmo32N8q8ZSQAIcU='"
 const getSessionCookieName = () => process.env.SESSION_COOKIE_NAME || SESSION_COOKIE_NAME_DEFAULT
 const getCsrfTokenCookieName = () => process.env.CSRF_TOKEN_COOKIE_NAME || CSRF_TOKEN_COOKIE_NAME_DEFAULT
-const getPlugIns = () => {
-  const hapiGapiPropertySettings = []
-  if (process.env.ANALYTICS_PRIMARY_PROPERTY) {
-    hapiGapiPropertySettings.push({
-      id: process.env.ANALYTICS_PRIMARY_PROPERTY,
-      hitTypes: ['pageview', 'event', 'ecommerce']
-    })
-  } else {
-    console.warn("ANALYTICS_PRIMARY_PROPERTY not set, so Google Analytics won't track this")
-  }
-  if (process.env.ANALYTICS_XGOV_PROPERTY) {
-    hapiGapiPropertySettings.push({
-      id: process.env.ANALYTICS_XGOV_PROPERTY,
-      hitTypes: ['pageview']
-    })
-  } else {
-    console.warn("ANALYTICS_XGOV_PROPERTY not set, so Google Analytics won't track this")
-  }
-
-  return [
-    Inert,
-    Vision,
-    Scooter,
-    {
-      plugin: Blankie,
-      options: {
-        /*
-         * This defines the content security policy - which is as restrictive as possible
-         * It must allow web-fonts from 'fonts.gstatic.com'
-         */
-        fontSrc: ['self', 'fonts.gstatic.com', 'data:'],
-        scriptSrc: [scriptHash],
-        generateNonces: true
-      }
-    },
-    {
-      plugin: Crumb,
-      options: {
-        key: getCsrfTokenCookieName(),
-        cookieOptions: {
-          isSecure: process.env.NODE_ENV !== 'development',
-          isHttpOnly: process.env.NODE_ENV !== 'development'
-        },
-        logUnauthorized: true
-      }
-    },
-    {
-      plugin: HapiGapi,
-      options: {
-        propertySettings: hapiGapiPropertySettings,
-        sessionIdProducer: request => (useSessionCookie(request) ? request.cache().getId() : null),
-        attributionProducer: async request => {
-          if (useSessionCookie(request)) {
-            const { attribution } = await request.cache().helpers.status.get()
-
-            if (attribution) {
-              return {
-                campaign: attribution[UTM.CAMPAIGN],
-                content: attribution[UTM.CONTENT],
-                medium: attribution[UTM.MEDIUM],
-                source: attribution[UTM.SOURCE],
-                term: attribution[UTM.TERM]
-              }
-            }
-          }
-          return {}
-        }
-      }
-    }
-  ]
-}
 
 /**
  * Adds the uri's used by the layout page to each relevant response
@@ -164,7 +83,7 @@ const layoutContextAmalgamation = (request, h) => {
 }
 
 const init = async () => {
-  await server.register(getPlugIns())
+  await server.register(getPlugIns(getSessionCookieName, getCsrfTokenCookieName))
   const viewPaths = [...new Set(find.fileSync(/\.njk$/, path.join(Dirname, './src/pages')).map(f => path.dirname(f)))]
 
   server.views({
